Add multiple UI related functions and types
This commit adds several new TypeScript files to the `/src/lib/ui` directory. These files include various asynchronous functions and types that are used for editing prompts, handling results, logging successes, and more.

- `editPrompt.ts` includes a function that uses the `editor` function from `@inquirer/prompts` to edit a prompt.
- `editResult.ts` allows the user to edit the result of a commit message if the `openInEditor` option is enabled.
- `generateAndReviewLoop.ts` generates and reviews a loop based on the provided options and input.
- `getUserReviewDecision.ts` prompts the user to make changes to the commit message and returns the user's decision.
- `handleResult.ts` handles the resulting commit message based on the mode provided.
- `helpers.ts` includes a function that checks if the mode is interactive and a constant named `SEPERATOR`.
- `logResult.ts` logs the result of a commit.
- `logSuccess.ts` logs a success message.

In addition to these additions, the `ui.ts` file in the `/src/lib` directory has been deleted.
