+++ src/lib/ui/helpers.ts
@@ -1,0 +1,8 @@
+import chalk from 'chalk'
+import { CommitOptions } from '../../commands/commit'
+
+export const isInteractive = (argv: CommitOptions) => {
+  return argv?.mode === 'interactive' || argv.interactive
+}
+
+export const SEPERATOR = chalk.blue('----------------')
