+++ src/lib/ui/logSuccess.ts
@@ -1,0 +1,5 @@
+import chalk from 'chalk';
+
+export const logSuccess = () => {
+  console.log(chalk.green(chalk.bold('\nAll set! 🦾🤖')))
+}
